+++ test/async_sign.tests.js
@@ -1,5 +1,6 @@
 var jwt = require('../index');
 var expect = require('chai').expect;
+var jws = require('jws');
 
 describe('signing a token asynchronously', function() {
 
@@ -13,6 +14,13 @@ describe('signing a token asynchronously', function() {
         expect(asyncToken).to.be.a('string');
         expect(asyncToken.split('.')).to.have.length(3);
         expect(asyncToken).to.equal(syncToken);
+        done();
+      });
+    });
+
+    it('should work', function (done) {
+      jwt.sign({abc: 1}, "secret", {}, function (err, res) {
+        expect(err).to.be.null();
         done();
       });
     });
@@ -37,6 +45,14 @@ describe('signing a token asynchronously', function() {
       jwt.sign('string', 'secret', {noTimestamp: true}, function (err) {
         expect(err).to.be.ok();
         expect(err).to.be.instanceof(Error);
+        done();
+      });
+    });
+
+    it('should not stringify the payload', function (done) {
+      jwt.sign('string', 'secret', {}, function (err, token) {
+        if (err) { return done(err); }
+        expect(jws.decode(token).payload).to.equal('string');
         done();
       });
     });